[tests] centralise stubs and reuse
