+++ tests/test_page_navigator_sequence.py
@@ -7,56 +7,15 @@ import pytest  # noqa: E402
 
 from sele_saisie_auto.encryption_utils import Credentials  # noqa: E402
 from sele_saisie_auto.navigation.page_navigator import PageNavigator  # noqa: E402
-
-
-class StubLoginHandler:
-    def __init__(self, log):
-        self.log = log
-
-    def connect_to_psatime(self, driver, key, login, pwd):
-        self.log.append("login")
-
-
-class StubDateEntryPage:
-    def __init__(self, log):
-        self.log = log
-
-    def navigate_from_home_to_date_entry_page(self, driver):
-        self.log.append("navigate")
-        return True
-
-    def process_date(self, driver, date):
-        self.log.append("process")
-
-
-class StubAdditionalInfoPage:
-    def __init__(self, log):
-        self.log = log
-
-    def navigate_from_work_schedule_to_additional_information_page(self, driver):
-        self.log.append("nav_add")
-
-    def submit_and_validate_additional_information(self, driver):
-        self.log.append("submit_add")
-
-    def save_draft_and_validate(self, driver):
-        self.log.append("save")
-
-
-class StubTimeSheetHelper:
-    def __init__(self, log):
-        self.log = log
-
-    def run(self, driver):
-        self.log.append("fill")
-
-
-class StubBrowserSession:
-    def __init__(self, log):
-        self.log = log
-
-    def go_to_default_content(self):
-        self.log.append("default")
+from tests.conftest import (  # noqa: E402
+    LoggedDummyAdditionalInfoPage as StubAdditionalInfoPage,
+)
+from tests.conftest import LoggedDummyDateEntryPage as StubDateEntryPage  # noqa: E402
+from tests.conftest import LoggedDummyLoginHandler as StubLoginHandler  # noqa: E402
+from tests.conftest import LoggedDummySession as StubBrowserSession  # noqa: E402
+from tests.conftest import (  # noqa: E402
+    LoggedDummyTimeSheetHelper as StubTimeSheetHelper,
+)
 
 
 def make_navigator():
